[cfe] Remove unknownFunction
The `unknownFunction` marker object was used to indicate a type of a
function that accepts an unspecified amount of arguments. The uses of
the marker object in the CFE are already replaced by
InvocationTargetType and its subtypes. This CL simply removes the
object itself, which is no longer in use.

Change-Id: I48fe53460d1f96f4fee0a91fb1572e219fbd1502
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/387702
Commit-Queue: Chloe Stefantsova <cstefantsova@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>