[linter] Handle erasure cycles in invalid_runtime_check_with_js_interop_types
Closes b/370036497

Enabling this lint in google3 lead to build failures when dealing
with generic interface types whose type parameter bounds point to
the interface type. By default, visitTypeParameterType does not
recurse on the bound, but EraseNonJSInteropTypes.visitTypeParameterType
does. By default, visitInterfaceType visits its type arguments, so
when we call super.visitInterfaceType and type.bound.accept(this),
this leads to a cycle, resulting in a stack overflow. Use a set
of visited types every time we perform the erasure to avoid this
cycle.

Change-Id: I2b874c43a751a8859c505b27715e2aa1d3682b6d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/387320
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Srujan Gaddam <srujzs@google.com>