[analysis_server] Return empty document highlights instead of null when no matches
This partly fixes https://github.com/Dart-Code/Dart-Code/issues/5294, in that it prevents VS Code falling back to a text search which gives confusing results.

However, because we don't produce any occurrences for record field names (because they are built on `Element`s which don't exist here), it also doesn't highlight the complete symbol name under the cursor like it would for other symbols (not does it highlight other references to the same record field in surrounding code). I'll follow up on the issue about this because it would also be nice to handle them.

Change-Id: I11f89d222d727cf73adf4a067ac00ca0cb5b1ebb
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/387961
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>