Don't suggest references to private enum fields
Some Dart code defines enums with private fields. This is useful to
indicate that the represented values might change in the future, and
should not considered to be exhaustive. Take this enum as an example:

```dart
enum ProtocolVersion {
  v1,
  v2,
  _upcoming,
}
```

When code in a different library is switching over that enum, it can't
use the private member:

```dart
import 'protocol_version.dart';

ProtocolImplementation findImplementation(ProtocolVersion version) {
  return switch (version) {
  	ProtocolVersion.v1 => LegacyProtocol(),
  	ProtocolVersion.v2 => NewProtocol(),
  	// Missing default case
  };
}
```

At the moment, the analysis server would suggest adding a reference to
`ProtocolVersion._upcoming` to fix the switch expression, which is
invalid. The only way to fix the switch is to add a default expression,
which this change introduces instead.

Change-Id: Id38dbf206a25c11dcc53af50e19ac5dd1287637e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/388005
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>