[analyzer] Change some if/else chains to switch statements.
These if/else chains were selecting a behavior based on the type of a
collection element. Changing them to switch statements, and making use
of `CollectionElementImpl` (which is a sealed type), ensures that if
more kinds of collection elements are added in the future, we won't
forget to update the switches.

Change-Id: Iedd5d14aa22b7690902868ba15d9127d79bcad27
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/388280
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>