[ddc] Reduce number of shards for firefox co19
This is a followup to
https://dart-review.googlesource.com/c/sdk/+/386900.

Since manually sharding the co19 tests on the firefox configuration
by subdirectories there have been no infrastructure failures on the
ddc-linux-firefox bot.

This change reduces the number of shards per each subdirectory
shard as an experiment to see if this will cause one of them to start
causing failures.

Change-Id: I1d8e0440134ade0c4c0be91e7ad8278b7bcacb01
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/388068
Reviewed-by: Nate Bosch <nbosch@google.com>
Commit-Queue: Nicholas Shahan <nshahan@google.com>