[cfe] Fix type field promotion on LHS of `??`.
The CFE was passing the original LHS expression to
`flowAnalysis.ifNullExpression_rightBegin`. This was incorrect; it
needs to pass the rewritten LHS expression. Most of the time this was
harmless, since in most cases:

- The expression on the LHS of `??` was not promotable, in which case
  passing the wrong expression didn't matter,

- Or the expression on the of `??` was a read of a local variable, in
  which case there was no problem, since local variable reads don't
  require rewriting.

However, in cases where the LHS of `??` is a read of a promotable
property, it does matter, since property reads are often rewritten
during inference.

Fixes https://github.com/dart-lang/sdk/issues/56874.

Bug: https://github.com/dart-lang/sdk/issues/56874
Change-Id: Ie42b96b0bc14faec38e8dbcb9e11534d4a1db0dc
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/389140
Commit-Queue: Paul Berry <paulberry@google.com>
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>