Remove most use cases of LintFilter.
The only time a `filter` was passed to LinterOptions was in
`benchmark.dart` and in tests of the filter option. But then
`benchmark.dart` just grabs it back out, so we can remove it from
LinterOptions.

It can then be removed from all of the reporters (DetailedReporter,
ReportFormatter, SimpleFormatter).

Remove unused `quiet` option from these reporters as well.

I suspect more can be removed from the reporters (again they are only
used in tests; no uses in the wild) but I'll save those for later.

Change-Id: I784666fb9c9cb42760fabec346aa5135af15e6fb
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/389403
Reviewed-by: Phil Quitslund <pquitslund@google.com>
Commit-Queue: Samuel Rawlins <srawlins@google.com>