Extract the interface Metadata
We can remove the methods in `Augmentable` that are now in `Metadata`
after this has landed and people can start using the new ones. Otherwise
we'll be in a state where new uses of the old are added as we're trying
to get this landed.

Change-Id: I711b3842040b1bc3875799708f01b542668c7a11
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/390561
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>