Elements. Store name2 in MethodFragment and similar.
Note that for setters we store the name without '='.
We don't use it yet, so this does not affect anything yet.

Note that for synthetic getter/setter fragments the name is `null`.
But we should not create such fragments at all.
Synthetic getter/setter elements are fine, but they should not have fragments.

Change-Id: I09bc07171d149572bea0a334b3f2e29ad4761b4c
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/390584
Commit-Queue: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>