Allow (e as dynamic).foo() in avoid_dynamic_calls
The `(e as dynamic)` pattern is the intended workaround for intentional
dynamic calls. The lint currently allows this for property reads like
`(e as dynamic).foo` but disallows this for method invocations like `(e
as dynamic).foo()` because the latter falls into a case checking if the
`node.function` is the dynamic expression (to handle cases like
`(e as dynamic).call()`) but there it is no longer checking the right
expression for whether it is a cast.

Pull out an `_isExplicitCast` method to check for casts wrapped in
parenthesis. Use the explicit cast as a signal to bail earlier when
checking method invocations.

Change-Id: I77b1c28b38474cea018b61a3d0ebe38ead82378d
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/390640
Commit-Queue: Nate Bosch <nbosch@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Auto-Submit: Nate Bosch <nbosch@google.com>