linter: Convert IOSinks to StringSink; use StringBuffer where possible
I think IOSink was used previously in these places because `Stdout`
implements IOSink. But IOSink also implements StringSink and only adds
some `add*` methods, and `close`, `flush`, `done`, and `encoding`,
which we don't use. Really, we are looking for a StringSink here.

Converting things to a StringSink means we can ditch CollectingSink,
and the `buffer` field on CollectingSink, in favor of a simple
StringBuffer, which also implements StringSink.

Change-Id: Id4a251424742493e37e9630f82c3e6c4a2b1d8bc
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/390665
Commit-Queue: Samuel Rawlins <srawlins@google.com>
Reviewed-by: Phil Quitslund <pquitslund@google.com>