[dap] Update to the latest version of the DAP spec
This is a result of running `dap/tool/generate_all.dart` to get the latest version of the spec. I'm doing this to get a new value `supportsANSIStyling` that was added recently that VS Code is already using and broke our ansi colors (it's now opt-in but because was just always supported).

This change does not fix that issue, I wanted to keep the spec updates separate.

In order to get a fix for VS Code users, I'm shipping some middleware that forces this on in the client, but we'll need to fix properly here for the benefit of other clients (or people who haven't updated their VS Code extensions?).

See https://github.com/Dart-Code/Dart-Code/issues/5302

Change-Id: Ie066b92da177bba5565d3ed952e5518912918f40
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/388700
Commit-Queue: Ben Konyi <bkonyi@google.com>
Reviewed-by: Ben Konyi <bkonyi@google.com>
Reviewed-by: Derek Xu <derekx@google.com>